requirements.txt
arrow==0.12.1
astroid==2.0.1
binaryornot==0.4.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
cookiecutter==1.6.0
Django==2.0.6
django-extensions==2.1.0
django-model-utils==3.1.2
et-xmlfile==1.0.1
fpdf==1.7.2
future==0.16.0
idna==2.7
isort==4.3.4
jdcal==1.4
Jinja2==2.10
jinja2-time==0.2.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
openpyxl==2.5.4
Pillow==5.2.0
poyo==0.4.1
pylint==2.0.1
python-dateutil==2.7.3
pytz==2018.4
requests==2.19.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
whichcraft==0.4.1
wrapt==1.10.11
xlrd==1.1.0
